kanban complete!update project entity!

+++ src/main/java/org/blank/projectmanagementsystem/domain/entity/Architecture.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import org.blank.projectmanagementsystem.domain.Enum.ArchitectureType;
 
 import java.io.Serializable;
+import java.util.Set;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
@@ -19,7 +20,7 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 public class Architecture implements Serializable {
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(unique = true, nullable = false, length = 25)
     private String name;
@@ -27,4 +28,5 @@ public class Architecture implements Serializable {
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private ArchitectureType type;
+
 }
