+++ src/main/java/org/blank/projectmanagementsystem/repository/SystemOutlineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Integer>{
+public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Long>{
     SystemOutline findByName(String systemOutlineName);
 }
