+++ src/main/java/org/blank/projectmanagementsystem/service/DeliverableService.java
@@ -5,8 +5,8 @@ import org.blank.projectmanagementsystem.domain.entity.Deliverable;
 public interface DeliverableService {
     Deliverable createDeliverable(Deliverable deliverable);
     Deliverable updateDeliverable(Deliverable deliverable);
-    void deleteDeliverable(int id);
-    Deliverable getDeliverable(int id);
+    void deleteDeliverable(Long id);
+    Deliverable getDeliverable(Long id);
